[vm] Move get_resource_group_member_from_metadata to aptos-types
This function only depends on aptos-types types (StructTag, Metadata,
get_metadata) but lived in aptos-vm/src/data_cache.rs, forcing
aptos-resource-viewer to depend on the entire aptos-vm crate.

Moving it to aptos-types::vm::module_metadata (where get_metadata already
lives) allows aptos-resource-viewer to drop its aptos-vm dependency. A
re-export is left in aptos-vm::data_cache for backwards compatibility.

This breaks the aptos-resource-viewer → aptos-vm transitive dependency
chain, benefiting all consumers of aptos-resource-viewer. Note that
aptos-consensus still sees aptos-vm transitively through aptos-scratchpad
→ aptos-executor-types, which is a deeper coupling to address separately.
