+++ aptos-move/aptos-resource-viewer/Cargo.toml
@@ -16,7 +16,6 @@ rust-version = { workspace = true }
 anyhow = { workspace = true }
 aptos-logger = { workspace = true }
 aptos-types = { workspace = true }
-aptos-vm = { workspace = true }
 aptos-vm-environment = { workspace = true }
 aptos-vm-logging = { workspace = true }
 move-binary-format = { workspace = true }
