+++ aptos-move/aptos-resource-viewer/src/lib.rs
@@ -8,8 +8,9 @@
 pub mod module_view;
 
 use crate::module_view::ModuleView;
-use aptos_types::state_store::StateView;
-use aptos_vm::data_cache::get_resource_group_member_from_metadata;
+use aptos_types::{
+    state_store::StateView, vm::module_metadata::get_resource_group_member_from_metadata,
+};
 use move_binary_format::CompiledModule;
 use move_core_types::{
     identifier::{IdentStr, Identifier},
